+++ hail/hail/src/is/hail/expr/ir/Simplify.scala
@@ -612,7 +612,7 @@ object Simplify {
         Some(xs.find(_._1 == idx).get._2)
 
       case TableCount(MatrixColsTable(ColumnCount(nCols))) =>
-        Some(I64(nCols))
+        Some(I64(nCols.toLong))
 
       case TableCount(PartitionCounts(counts)) =>
         Some(I64(counts.sum))
@@ -645,7 +645,7 @@ object Simplify {
         Some(TableCount(child))
 
       case TableCount(TableRange(n, _)) =>
-        Some(I64(n))
+        Some(I64(n.toLong))
 
       case TableCount(TableParallelize(rowsAndGlobal, _)) =>
         Some(Cast(ArrayLen(GetField(rowsAndGlobal, "rows")), TInt64))
