+++ hail/hail/test/src/is/hail/annotations/UnsafeSuite.scala
@@ -217,7 +217,7 @@ class UnsafeSuite extends HailSuite with ScalaCheckDrivenPropertyChecks {
 
       // test addAnnotation
       region.clear()
-      region.allocate(1, n): Unit // preallocate
+      region.allocate(1, n.toLong): Unit // preallocate
 
       val offset = pt.unstagedStoreJavaObject(sm, a, region)
 
@@ -230,7 +230,7 @@ class UnsafeSuite extends HailSuite with ScalaCheckDrivenPropertyChecks {
 
       // test addAnnotation from ur
       region2.clear()
-      region2.allocate(1, n2): Unit // preallocate
+      region2.allocate(1, n2.toLong): Unit // preallocate
       val offset2 = pt.unstagedStoreJavaObject(sm, ur, region2)
 
       val ur2 = UnsafeRow.read(pt, region2, offset2)
@@ -238,7 +238,7 @@ class UnsafeSuite extends HailSuite with ScalaCheckDrivenPropertyChecks {
 
       // test addRegionValue
       region2.clear()
-      region2.allocate(1, n2): Unit // preallocate
+      region2.allocate(1, n2.toLong): Unit // preallocate
       rvb2.start(pt)
       rvb2.addRegionValue(pt, region, offset)
       val offset3 = rvb2.end()
@@ -250,7 +250,7 @@ class UnsafeSuite extends HailSuite with ScalaCheckDrivenPropertyChecks {
         case t: TStruct =>
           val ps = pt.asInstanceOf[PStruct]
           region2.clear()
-          region2.allocate(1, n): Unit // preallocate
+          region2.allocate(1, n.toLong): Unit // preallocate
           val offset4 =
             ps.unstagedStoreJavaObject(sm, Row.fromSeq(a.asInstanceOf[Row].toSeq), region2)
           val ur4 = new UnsafeRow(ps, region2, offset4)
